requirements.txt
argon2-cffi==20.1.0
astroid==2.4.2
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.3
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
pylint==2.5.3
pylint-django==2.1.0
pylint-plugin-utils==0.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.1.0
wincertstore==0.2
wrapt==1.12.1
